+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 - (VANAGON-171) Add Ubuntu 18.04 aarch64 defaults
 - (VANAGON-166) Add RedHat 8 FIPS defaults
 - (VANAGON-166) Do not undefine `__debug_package` on EL 8 FIPS
+- (maint) Update vmpooler URL
 
 ## [0.22.0] - released 2021-07-06
 ### Added